requirements.txt
fastapi
uvicorn
langchain
unstructured
pinecone-client
openai
pydantic
pymongo
requests
beautifulsoup4
